Update timer docu

+++ modules/rtos/src/cobc/rtos/rtems/clock.cpp
@@ -25,7 +25,8 @@ cobc::rtos::Clock::now()
 	rtems_clock_get(RTEMS_CLOCK_GET_TICKS_SINCE_BOOT, &ticks_since_boot);
 
 	// convert to microseconds
-	uint64_t us = ticks_since_boot * us_per_tick;
+	uint64_t us = static_cast<uint64_t>(ticks_since_boot) *
+	              static_cast<uint64_t>(us_per_tick);
 
 	return cobc::time::TimePoint(us);
 #endif
